Create a new workflow for the current release
The current release is meant to contain the artifacts that should be used
in the validation pipeline at the current moment. It differs from "latest",
which is built automatically and might contain changes that we do not want
to be used for now.

To create a new current release, the new workflow must be triggered manually
with the git sha of the commit from where we want to build the artifacts.

The artifacts will be available on the following permalink:
```
https://github.com/{owner}/{repo}/releases/download/{release-name}/{artifact-name}
```