Refactor active calculation
Refactor the active memory footprint calculation so that the core logic can be used in the ambient calculation as well, to optimise the way we evaluate nodes under multiple combinations of user styles.

This change also removes some code that was unused after changes were made to the ambient calculation.