Avoid stack overflow in AmbientMemoryFootprintCalculator
On some watch faces, there can be too many permutations of style configurations to compute, leading to a stack overflow or to a very long time when evaluating the memory footprint in Ambient.

With this change, instead of generating all the possible configurations of user style configurations, we only generate combinations that are necessary, for resources that cannot be evaluated independently.