Remove residual logs
Remove a log that was left from a previous change and which adds to much spam to the output. This is important especially for the --report evaluation mode, where the tool should only print a JSON that can be consumed by other tools. Add a test that checks if the --report evaluation mode prints something that can be parsed as JSON to make sure no future residual logs break this behaviour.