Simplify AndroidManifest parsing (#69)
Reduce the size of the binary by replacing and reduce the libraries used to parse the android manifest of an input package. For APKs, the library is replaced with a more lightweight one, while for AABs, the parsing is refactored to work on the Proto XML nodes directly, removing the need for adding the bundletool dependency.