Clean up temp dirs after validation
Delete residual xsd files after parsing the XSD schemas. Before, these
files were left on the disk because the Java built-in deleteOnExit does
not delete non-empty directories.

This change also simlifies the code by removing the ResourceManager
class and parsing all the schemas on initialization, deleting the temp
files as soon as possible.